requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-telegram-bot==21.9
anthropic==0.42.0
python-dotenv==1.0.1
pydantic==2.10.4
